+++ rust/dbn/src/publishers.rs
@@ -105,10 +105,18 @@ pub enum Venue {
     Aspi = 46,
     /// Databento US Equities - Consolidated
     Equs = 47,
+    /// ICE Futures US
+    Ifus = 48,
+    /// ICE Futures Europe (Financials)
+    Ifll = 49,
+    /// Eurex Exchange
+    Xeur = 50,
+    /// European Energy Exchange
+    Xeer = 51,
 }
 
 /// The number of Venue variants.
-pub const VENUE_COUNT: usize = 47;
+pub const VENUE_COUNT: usize = 51;
 
 impl Venue {
     /// Convert a Venue to its `str` representation.
@@ -161,6 +169,10 @@ impl Venue {
             Self::Asmt => "ASMT",
             Self::Aspi => "ASPI",
             Self::Equs => "EQUS",
+            Self::Ifus => "IFUS",
+            Self::Ifll => "IFLL",
+            Self::Xeur => "XEUR",
+            Self::Xeer => "XEER",
         }
     }
 }
@@ -229,6 +241,10 @@ impl std::str::FromStr for Venue {
             "ASMT" => Ok(Self::Asmt),
             "ASPI" => Ok(Self::Aspi),
             "EQUS" => Ok(Self::Equs),
+            "IFUS" => Ok(Self::Ifus),
+            "IFLL" => Ok(Self::Ifll),
+            "XEUR" => Ok(Self::Xeur),
+            "XEER" => Ok(Self::Xeer),
             _ => Err(Error::conversion::<Self>(s)),
         }
     }
@@ -311,10 +327,18 @@ pub enum Dataset {
     XnysTradesbbo = 34,
     /// Databento US Equities Mini
     EqusMini = 35,
+    /// ICE Futures US iMpact
+    IfusImpact = 36,
+    /// ICE Futures Europe (Financials) iMpact
+    IfllImpact = 37,
+    /// Eurex EOBI
+    XeurEobi = 38,
+    /// European Energy Exchange EOBI
+    XeerEobi = 39,
 }
 
 /// The number of Dataset variants.
-pub const DATASET_COUNT: usize = 35;
+pub const DATASET_COUNT: usize = 39;
 
 impl Dataset {
     /// Convert a Dataset to its `str` representation.
@@ -357,6 +381,10 @@ impl Dataset {
             Self::XcisTradesbbo => "XCIS.TRADESBBO",
             Self::XnysTradesbbo => "XNYS.TRADESBBO",
             Self::EqusMini => "EQUS.MINI",
+            Self::IfusImpact => "IFUS.IMPACT",
+            Self::IfllImpact => "IFLL.IMPACT",
+            Self::XeurEobi => "XEUR.EOBI",
+            Self::XeerEobi => "XEER.EOBI",
         }
     }
 }
@@ -415,6 +443,10 @@ impl std::str::FromStr for Dataset {
             "XCIS.TRADESBBO" => Ok(Self::XcisTradesbbo),
             "XNYS.TRADESBBO" => Ok(Self::XnysTradesbbo),
             "EQUS.MINI" => Ok(Self::EqusMini),
+            "IFUS.IMPACT" => Ok(Self::IfusImpact),
+            "IFLL.IMPACT" => Ok(Self::IfllImpact),
+            "XEUR.EOBI" => Ok(Self::XeurEobi),
+            "XEER.EOBI" => Ok(Self::XeerEobi),
             _ => Err(Error::conversion::<Self>(s)),
         }
     }
@@ -617,10 +649,22 @@ pub enum Publisher {
     EqusMiniEqus = 95,
     /// NYSE Trades - Consolidated
     XnysTradesEqus = 96,
+    /// ICE Futures US
+    IfusImpactIfus = 97,
+    /// ICE Futures US - Off-Market Trades
+    IfusImpactXoff = 98,
+    /// ICE Futures Europe (Financials)
+    IfllImpactIfll = 99,
+    /// ICE Futures Europe (Financials) - Off-Market Trades
+    IfllImpactXoff = 100,
+    /// Eurex EOBI
+    XeurEobiXeur = 101,
+    /// European Energy Exchange EOBI
+    XeerEobiXeer = 102,
 }
 
 /// The number of Publisher variants.
-pub const PUBLISHER_COUNT: usize = 96;
+pub const PUBLISHER_COUNT: usize = 102;
 
 impl Publisher {
     /// Convert a Publisher to its `str` representation.
@@ -722,6 +766,12 @@ impl Publisher {
             Self::EqusAllEqus => "EQUS.ALL.EQUS",
             Self::EqusMiniEqus => "EQUS.MINI.EQUS",
             Self::XnysTradesEqus => "XNYS.TRADES.EQUS",
+            Self::IfusImpactIfus => "IFUS.IMPACT.IFUS",
+            Self::IfusImpactXoff => "IFUS.IMPACT.XOFF",
+            Self::IfllImpactIfll => "IFLL.IMPACT.IFLL",
+            Self::IfllImpactXoff => "IFLL.IMPACT.XOFF",
+            Self::XeurEobiXeur => "XEUR.EOBI.XEUR",
+            Self::XeerEobiXeer => "XEER.EOBI.XEER",
         }
     }
 
@@ -824,6 +874,12 @@ impl Publisher {
             Self::EqusAllEqus => Venue::Equs,
             Self::EqusMiniEqus => Venue::Equs,
             Self::XnysTradesEqus => Venue::Equs,
+            Self::IfusImpactIfus => Venue::Ifus,
+            Self::IfusImpactXoff => Venue::Xoff,
+            Self::IfllImpactIfll => Venue::Ifll,
+            Self::IfllImpactXoff => Venue::Xoff,
+            Self::XeurEobiXeur => Venue::Xeur,
+            Self::XeerEobiXeer => Venue::Xeer,
         }
     }
 
@@ -926,6 +982,12 @@ impl Publisher {
             Self::EqusAllEqus => Dataset::EqusAll,
             Self::EqusMiniEqus => Dataset::EqusMini,
             Self::XnysTradesEqus => Dataset::XnysTrades,
+            Self::IfusImpactIfus => Dataset::IfusImpact,
+            Self::IfusImpactXoff => Dataset::IfusImpact,
+            Self::IfllImpactIfll => Dataset::IfllImpact,
+            Self::IfllImpactXoff => Dataset::IfllImpact,
+            Self::XeurEobiXeur => Dataset::XeurEobi,
+            Self::XeerEobiXeer => Dataset::XeerEobi,
         }
     }
 
@@ -1030,6 +1092,12 @@ impl Publisher {
             (Dataset::EqusAll, Venue::Equs) => Ok(Self::EqusAllEqus),
             (Dataset::EqusMini, Venue::Equs) => Ok(Self::EqusMiniEqus),
             (Dataset::XnysTrades, Venue::Equs) => Ok(Self::XnysTradesEqus),
+            (Dataset::IfusImpact, Venue::Ifus) => Ok(Self::IfusImpactIfus),
+            (Dataset::IfusImpact, Venue::Xoff) => Ok(Self::IfusImpactXoff),
+            (Dataset::IfllImpact, Venue::Ifll) => Ok(Self::IfllImpactIfll),
+            (Dataset::IfllImpact, Venue::Xoff) => Ok(Self::IfllImpactXoff),
+            (Dataset::XeurEobi, Venue::Xeur) => Ok(Self::XeurEobiXeur),
+            (Dataset::XeerEobi, Venue::Xeer) => Ok(Self::XeerEobiXeer),
             _ => Err(Error::conversion::<Self>(format!("({dataset}, {venue})"))),
         }
     }
@@ -1148,6 +1216,12 @@ impl std::str::FromStr for Publisher {
             "EQUS.ALL.EQUS" => Ok(Self::EqusAllEqus),
             "EQUS.MINI.EQUS" => Ok(Self::EqusMiniEqus),
             "XNYS.TRADES.EQUS" => Ok(Self::XnysTradesEqus),
+            "IFUS.IMPACT.IFUS" => Ok(Self::IfusImpactIfus),
+            "IFUS.IMPACT.XOFF" => Ok(Self::IfusImpactXoff),
+            "IFLL.IMPACT.IFLL" => Ok(Self::IfllImpactIfll),
+            "IFLL.IMPACT.XOFF" => Ok(Self::IfllImpactXoff),
+            "XEUR.EOBI.XEUR" => Ok(Self::XeurEobiXeur),
+            "XEER.EOBI.XEER" => Ok(Self::XeerEobiXeer),
             _ => Err(Error::conversion::<Self>(s)),
         }
     }